Create basic test endpoint class
